requirements.txt
PyYAML==6.0.1
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2 